Put message in types so we can use it on the frontend
